Turn hasher-priv into a daemon
All the privileged operations are now performed by a daemonised system
service.  The daemon accepts IPC commands via a Unix domain socket,
fulfills the request and sends back a response.
The kernel provides a reliable way of obtaining the sender's credentials,
which are then verified.

When contacted by a client of user X, the daemon creates a separate
session process that handles requests only from user X.  This session
process ends after a certain period of inactivity.

hasher-privd is now installed without set-UID bit, reducing the attack
surface and facilitating NO_NEW_PRIVS environments.

+++ hasher-priv/pass.c
@@ -141,7 +141,7 @@ fd_recv(int ctl, int *fds, unsigned int n_fds, char *data, size_t data_len)
 	size_t recv_len = cmsg->cmsg_len - CMSG_LEN(0);
 	if (recv_len != clen) {
 		error_msg("SCM_RIGHTS: expected size %lu, got %lu",
-			  clen, recv_len);
+			  (unsigned long) clen, (unsigned long) recv_len);
 		return -1;
 	}
